Added fade in/out, basic grid system
Added a fade out and in between scenes, and a bool to stop player movement during the fadeout
Moved scene changing functionality out of the player controller and into a separate script and made a scene changer prefab.
Added a basic grid system - plenty to be improved on there and there's probably better ways to approach the problem but it's a good(ish) starting point.