Finished setting up tutorial
Basic tutorial -takes away some player agency but runs them through all the necessary stages - tilling, planting, watering and harvesting, as well as the codependency system of feeding crops to energise the spaceship and using the spaceship to increase stamina.

Code needs some clearing up as does the whole project.

A couple of tutorial stages are incomplete as branch needs to be merged with recent changes on main